rename recovery data with default column names

This function renames the recovery data with default column names.

renameData(recoveryData, xname, yname = NULL, timename, markAreaName)

Arguments

recoveryData
data.frame with dead recoveries, containing information on breeding origin, geographical coordinates of recovery and age at death

xname
character, name of column containing information on longitude, specify this in one-dimensional case

yname
character, name of column containing information on latitude, defaults to NULL, do not specify in one-dimensional case, but use xname instead

timename
character, name of column containing information on age at death

markAreaName
character, name of column containing information on breeding origin

Value

recoveryData with default column names

Examples

{
    renameData(recoveryData = robinsInit$recoveryData,
               xname        = "recLon",
               yname        = "recLat",
               timename     = "age",
               markAreaName = "markArea")
}

#>     markArea latitude   longitude age
#> 2       late 54.30000 13.68333333   1
#> 3       late 52.59278 13.39888889   2
#> 5       late 55.70889 14.00444444   1
#> 11      late 53.76667 12.50000000   1
#> 22      late 53.09833 13.57750000   1
#> 27      late 53.03333 14.00000000   1
#> 30      late 46.80000 12.85000000   2
#> 35     early 45.75000 12.31666667   1
#> 37      late 51.76667  8.31666667   2
#> 39      late 48.36667  7.70000000   2
#> 40      late 48.68333  6.30000000   3
#> 41      late 50.93333  6.95000000   1
#> 42    medium 46.68333 -0.08333333   1
#> 43     early 51.43333  4.11666667   1
#> 45     early 50.68333  4.36666667   2
#> 46    medium 43.96667 -0.16666667   1
#> 47    medium 53.08333  8.83333333   1
#> 48     early 47.48333 -0.53333333   2
#> 49     early 46.28333 -0.15000000   1
#> 55      late 50.58333  2.40000000   2
#> 56     early 46.81667 -2.00000000   1
#> 57      late 46.18333  4.65000000   1
#> 58    medium 45.58333  4.76666667   1
#> 59     early 44.77417 -0.54083333   3
#> 60      late 43.73333 -0.01666667   1
#> 62    medium 46.85000 -1.86666667   2
#> 63      late 54.31667  8.95000000   2
#> 66    medium 51.15000  5.60000000   1
#> 68      late 53.37750  5.29416667   2
#> 71      late 49.05000  6.60000000   1
#> 75      late 52.43333 10.80000000   1
#> 78    medium 45.35000 -0.83333333   2
#> 79      late 52.26667 10.06666667   1
#> 81     early 34.13333  2.05000000   1
#> 82      late 38.08333 -2.63333333   2
#> 83     early 41.10000  0.93333333   3
#> 84     early 36.18333  5.40000000   1
#> 85     early 36.45000  4.25000000   1
#> 87     early 34.41667 -4.65000000   2
#> 88     early 37.58333 -4.08333333   1
#> 89    medium 36.71667  4.03333333   1
#> 90     early 34.00000 -5.50000000   1
#> 91      late 36.73333  4.08333333   1
#> 94      late 43.66667  3.76666667   2
#> 95     early 37.30000 -7.83333333   2
#> 96     early 37.50000 -8.50000000   1
#> 98     early 40.09333 -6.40277778   3
#> 100   medium 43.03333  3.05000000   1