raw real-world data for robins

A dataset containing the raw data of robins marked at the ringing station Greifswalder Oie, Germany, during autumn migration and dead recoveries between November and February in Europe and Africa.

data(robinsInit)

Format

A list of objects.

markData

data.frame with names of the area where the marking took place (markArea) and number of marked individuals per area

recoveryData

data.frame with one row for every dead recovery. Columns are names of the area where the marking took place (markArea), geographical coordinates (recLat and recLon) and age at death (age)

window

object of class owin containing the boundaries of the world map in a rectangle around the data points

observationTime

numeric, length of observation time. One length for all individuals in the dataset.

References

The data is provided by the Verein Jordsand e.V. and the Beringungszentrale Hiddensee.

Examples

data(robinsInit)
str(robinsInit)

#> List of 4
#>  $ markData       :'data.frame':	3 obs. of  2 variables:
#>   ..$ markArea  : chr [1:3] "early" "medium" "late"
#>   ..$ markedInds: num [1:3] 34177 32389 31833
#>  $ recoveryData   :'data.frame':	48 obs. of  4 variables:
#>   ..$ markArea: Factor w/ 3 levels "early","late",..: 2 2 2 2 2 2 2 1 2 2 ...
#>   ..$ recLat  : num [1:48] 54.3 52.6 55.7 53.8 53.1 ...
#>   ..$ recLon  : num [1:48] 13.7 13.4 14 12.5 13.6 ...
#>   ..$ age     : int [1:48] 1 2 1 1 1 1 2 1 2 2 ...
#>  $ window         :List of 5
#>   ..$ type  : chr "polygonal"
#>   ..$ xrange: num [1:2] -15 30
#>   ..$ yrange: num [1:2] 30 65
#>   ..$ bdry  :List of 18
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:20] -14.5 -14.5 -14.5 -14.2 -14 ...
#>   .. .. ..$ y: num [1:20] 63.6 63.7 63.7 63.8 63.9 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:1665] 30 29.8 29.7 29.6 29.5 ...
#>   .. .. ..$ y: num [1:1665] 32.1 32 32 31.9 31.9 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:77] 19 18.9 18.7 18.5 18.4 ...
#>   .. .. ..$ y: num [1:77] 60.7 60.8 60.9 61 61 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:51] -6.47 -6.43 -6.25 -6.12 -6.04 ...
#>   .. .. ..$ y: num [1:51] 60.8 60.8 60.9 61.1 61.3 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:39] -1.055 -0.999 -0.789 -0.768 -0.607 ...
#>   .. .. ..$ y: num [1:39] 59.2 59.2 59.4 59.4 59.5 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:152] 15.7 15.7 15.6 15.6 15.6 ...
#>   .. .. ..$ y: num [1:152] 55 55 55.3 55.5 55.5 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:5] 9.73 9.77 9.78 9.84 9.73
#>   .. .. ..$ y: num [1:5] 58.3 58.3 58.3 58.3 58.3
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:13] 17.3 17.3 17.3 17.3 17.3 ...
#>   .. .. ..$ y: num [1:13] 58 58 58 58.1 58.1 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:6] 11.5 11.5 11.5 11.5 11.5 ...
#>   .. .. ..$ y: num [1:6] 56.7 56.7 56.8 56.8 56.7 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:13] -5.3 -5.33 -5.3 -5.29 -5.29 ...
#>   .. .. ..$ y: num [1:13] 53.1 53.2 53.3 53.4 53.6 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:5] -5.34 -5.29 -5.27 -5.25 -5.37
#>   .. .. ..$ y: num [1:5] 52.6 52.8 52.6 52.6 52.5
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:3] -0.431 -0.255 -0.292
#>   .. .. ..$ y: num [1:3] 50 50.1 50
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:3] 13.1 13.1 13.1
#>   .. .. ..$ y: num [1:3] 44.6 44.7 44.6
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:48] 15.5 15.4 15.3 15.2 15.2 ...
#>   .. .. ..$ y: num [1:48] 42.7 42.7 42.7 42.7 42.7 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:14] 8.62 8.75 8.79 8.87 8.9 ...
#>   .. .. ..$ y: num [1:14] 43.4 43.5 43.6 43.7 43.7 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:45] 18.7 18.6 18.6 18.5 18.3 ...
#>   .. .. ..$ y: num [1:45] 41 41.1 41.1 41.2 41.4 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:20] 25.1 25 24.9 24.8 24.7 ...
#>   .. .. ..$ y: num [1:20] 38.6 38.8 38.8 38.9 39.1 ...
#>   .. ..$ :List of 2
#>   .. .. ..$ x: num [1:18] -3.53 -3.61 -3.67 -3.7 -3.88 ...
#>   .. .. ..$ y: num [1:18] 36 36 36 36 36 ...
#>   ..$ units :List of 3
#>   .. ..$ singular  : chr "unit"
#>   .. ..$ plural    : chr "units"
#>   .. ..$ multiplier: num 1
#>   .. ..- attr(*, "class")= chr "unitname"
#>   ..- attr(*, "class")= chr "owin"
#>  $ observationTime: num 5