CONSURE: Spatially continuous survival, use of space and recovery probability estimates

The CONSURE package provides functions to perform the continuous and the combined approach from the dissertation of Saskia Schirmer. It can work with both, simulated and real-world data.

Details

Nearly all functions are based on a markRecaptureObject.

A spatial point pattern can be simulated by given survival, migratory connectivity and recovery probability using simContin.

Visualizing raw data

Spatial point patterns of raw recoveries can be visualized by plotRawRecoveries. The age distribution can be visualized by plotAgeDistribution.

Continuous functions

The continuous estimation approach first estimates the density of the point pattern by estKDE. Then, survival can be estimated by estS and finally migratory connectivity by estM and a constant recovery probability by estR.

All estimates can be plotted by the appropriate function: plotKDE, plotS, plotM. The R^2 values of the linear model used to fit survival, migratory connectivity and recovery probability can be plotted by plotGOFofLM.

The number of recovered individuals per breeding area can be summarized by recIndsFunc.

par_grid creates a grid containing the values of a specific function on the grid.